Moving search-button into the nhsuk-frontend components folder
